+++ nodalia-cards-2.1.3-alpha.3.js
@@ -1,2 +1,2 @@
-import "./nodalia-cards-2.2.0-alpha.2.js";
-if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.2.0-alpha.2","contentSha256_12":"821029e6b1cf","file":"nodalia-cards-2.1.3-alpha.3.js","targetFile":"nodalia-cards-2.2.0-alpha.2.js","fallbackFile":"nodalia-cards.js"};}
+import "./nodalia-cards-2.2.0-alpha.3.js";
+if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.2.0-alpha.3","contentSha256_12":"c7cc53b1d742","file":"nodalia-cards-2.1.3-alpha.3.js","targetFile":"nodalia-cards-2.2.0-alpha.3.js","fallbackFile":"nodalia-cards.js"};}