+++ nodalia-cards.manifest.js
@@ -1,6 +1,6 @@
 export default {
-  "pkgVersion": "2.2.0-alpha.2",
-  "contentSha256_12": "821029e6b1cf",
+  "pkgVersion": "2.2.0-alpha.3",
+  "contentSha256_12": "c7cc53b1d742",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
   "hacsFile": "nodalia-cards.js",
@@ -8,16 +8,16 @@ export default {
     "nodalia-cards-2.1.3-alpha.2.js",
     "nodalia-cards-2.1.3-alpha.3.js"
   ],
-  "splitCoreFile": "nodalia-cards-core-2.2.0-alpha.2.js",
+  "splitCoreFile": "nodalia-cards-core-2.2.0-alpha.3.js",
   "splitCoreSha256_12": "ca962925538d",
-  "splitSuiteFile": "nodalia-cards-suite-2.2.0-alpha.2.js",
-  "splitSuiteSha256_12": "1b1f58e95576",
-  "editorFile": "nodalia-cards-editor-2.2.0-alpha.2.js",
+  "splitSuiteFile": "nodalia-cards-suite-2.2.0-alpha.3.js",
+  "splitSuiteSha256_12": "142b8670d039",
+  "editorFile": "nodalia-cards-editor-2.2.0-alpha.3.js",
   "editorSha256_12": "b22dc56a3814"
 };
-export const pkgVersion = "2.2.0-alpha.2";
-export const contentSha256_12 = "821029e6b1cf";
+export const pkgVersion = "2.2.0-alpha.3";
+export const contentSha256_12 = "c7cc53b1d742";
 export const file = "nodalia-cards.bundle.js";
-export const splitCoreFile = "nodalia-cards-core-2.2.0-alpha.2.js";
-export const splitSuiteFile = "nodalia-cards-suite-2.2.0-alpha.2.js";
-export const editorFile = "nodalia-cards-editor-2.2.0-alpha.2.js";
+export const splitCoreFile = "nodalia-cards-core-2.2.0-alpha.3.js";
+export const splitSuiteFile = "nodalia-cards-suite-2.2.0-alpha.3.js";
+export const editorFile = "nodalia-cards-editor-2.2.0-alpha.3.js";